Add configuration, split down into two event categories
