+++ src/main/kotlin/org/hyacinthbots/lilybot/database/entities/Config.kt
@@ -70,5 +70,9 @@ data class ModerationConfigData(
 @Serializable
 data class UtilityConfigData(
 	val guildId: Snowflake,
-	val utilityLogChannel: Snowflake?
+	val utilityLogChannel: Snowflake?,
+	val logChannelUpdates: Boolean,
+	val logEventUpdates: Boolean,
+	val logInviteUpdates: Boolean,
+	val logRoleUpdates: Boolean
 )
