+++ src/main/kotlin/org/hyacinthbots/lilybot/extensions/utils/events/UtilityEvents.kt
@@ -1,11 +1,10 @@
-package org.hyacinthbots.lilybot.extensions.logging.events
+package org.hyacinthbots.lilybot.extensions.utils.events
 
 import dev.kord.common.entity.ChannelType
 import dev.kord.common.entity.ForumTag
 import dev.kord.common.entity.GuildScheduledEventPrivacyLevel
 import dev.kord.common.entity.optional.value
 import dev.kord.core.behavior.GuildBehavior
-import dev.kord.core.behavior.UserBehavior
 import dev.kord.core.behavior.channel.asChannelOfOrNull
 import dev.kord.core.behavior.channel.createEmbed
 import dev.kord.core.behavior.channel.createMessage
@@ -17,15 +16,12 @@ import dev.kord.core.event.channel.ChannelDeleteEvent
 import dev.kord.core.event.channel.ChannelUpdateEvent
 import dev.kord.core.event.channel.thread.ThreadChannelCreateEvent
 import dev.kord.core.event.channel.thread.ThreadChannelDeleteEvent
-import dev.kord.core.event.guild.BanAddEvent
-import dev.kord.core.event.guild.BanRemoveEvent
 import dev.kord.core.event.guild.GuildScheduledEventCreateEvent
 import dev.kord.core.event.guild.GuildScheduledEventDeleteEvent
 import dev.kord.core.event.guild.GuildScheduledEventEvent
 import dev.kord.core.event.guild.GuildScheduledEventUpdateEvent
 import dev.kord.core.event.guild.InviteCreateEvent
 import dev.kord.core.event.guild.InviteDeleteEvent
-import dev.kord.core.event.guild.MemberUpdateEvent
 import dev.kord.core.event.role.RoleCreateEvent
 import dev.kord.core.event.role.RoleDeleteEvent
 import dev.kord.core.event.role.RoleUpdateEvent
@@ -39,18 +35,14 @@ import dev.kordex.core.extensions.Extension
 import dev.kordex.core.extensions.event
 import dev.kordex.core.time.TimestampType
 import dev.kordex.core.time.toDiscord
-import dev.kordex.core.utils.timeoutUntil
 import kotlinx.datetime.Clock
-import org.hyacinthbots.lilybot.database.collections.ModerationActionCollection
+import org.hyacinthbots.lilybot.database.collections.UtilityConfigCollection
 import org.hyacinthbots.lilybot.extensions.config.ConfigOptions
-import org.hyacinthbots.lilybot.extensions.moderation.ModerationAction
 import org.hyacinthbots.lilybot.utils.afterDot
-import org.hyacinthbots.lilybot.utils.baseModerationEmbed
-import org.hyacinthbots.lilybot.utils.dmNotificationStatusEmbedField
 import org.hyacinthbots.lilybot.utils.formatPermissionSet
 import org.hyacinthbots.lilybot.utils.getLoggingChannelWithPerms
 import org.hyacinthbots.lilybot.utils.ifNullOrEmpty
-import org.hyacinthbots.lilybot.utils.interval
+import kotlin.collections.forEach
 
 /** A String identifier to use for the permission map to get allowed permissions. */
 private const val ALLOWED = "Allowed"
@@ -58,164 +50,10 @@ private const val ALLOWED = "Allowed"
 /** A String identifier to use for the permission map to get denied permissions. */
 private const val DENIED = "Denied"
 
-class ModerationEvents : Extension() {
-	override val name: String = "moderation-events"
+class UtilityEvents : Extension() {
+	override val name: String = "utility-events"
 
 	override suspend fun setup() {
-		event<BanAddEvent> {
-			check { anyGuild() }
-			action {
-				// If the ban doesn't exist then... ????
-				event.getBanOrNull() ?: return@action
-				var existingAction = ModerationActionCollection().getAction(
-					ModerationAction.BAN, event.guildId, event.user.id
-				)
-				if (existingAction == null) {
-					existingAction = ModerationActionCollection().getAction(
-						ModerationAction.SOFT_BAN, event.guildId, event.user.id
-					)
-				}
-
-				if (existingAction == null) {
-					existingAction = ModerationActionCollection().getAction(
-						ModerationAction.TEMP_BAN, event.guildId, event.user.id
-					)
-				}
-
-				if (existingAction != null && existingAction.targetUserId != event.user.id) {
-					return@action
-				}
-
-				if (existingAction != null) {
-					getLoggingChannelWithPerms(ConfigOptions.ACTION_LOG, event.getGuild())?.createEmbed {
-						title = when (existingAction.actionType) {
-							ModerationAction.BAN -> "Banned a user"
-							ModerationAction.SOFT_BAN -> "Soft-banned a user"
-							ModerationAction.TEMP_BAN -> "Temp-banned a user"
-							else -> null // Theoretically this should never occur but the compiler cries otherwise
-						}
-
-						description = "${event.user.mention} has been ${
-							if (existingAction.data.reason?.contains("quick ban", true) == false) {
-								when (existingAction.actionType) {
-									ModerationAction.BAN -> "banned"
-									ModerationAction.SOFT_BAN -> "soft-banned"
-									ModerationAction.TEMP_BAN -> "temporarily banned"
-									else -> null // Again should theoretically never occur, but compiler
-								}
-							} else {
-								when (existingAction.actionType) {
-									ModerationAction.BAN -> existingAction.data.reason
-									ModerationAction.SOFT_BAN -> existingAction.data.reason
-									else -> null
-								}
-							}
-						}"
-						baseModerationEmbed(
-							existingAction.data.reason,
-							event.user,
-							UserBehavior(existingAction.data.actioner!!, kord)
-						)
-						image = existingAction.data.imageUrl
-						dmNotificationStatusEmbedField(existingAction.data.dmOutcome, existingAction.data.dmOverride)
-						timestamp = Clock.System.now()
-						if (existingAction.data.deletedMessages != null) {
-							field {
-								name = "Days of messages deleted"
-								value =
-									if (existingAction.actionType == ModerationAction.SOFT_BAN &&
-										existingAction.data.deletedMessages == 0
-									) {
-										"3"
-									} else {
-										existingAction.data.deletedMessages.toString()
-									}
-								inline = false
-							}
-						}
-						if (existingAction.data.timeData != null) {
-							field {
-								name = "Duration:"
-								value =
-									existingAction.data.timeData.durationInst?.toDiscord(TimestampType.Default) +
-										" (${existingAction.data.timeData.durationDtp?.interval()})"
-							}
-						}
-					}
-
-					ModerationActionCollection().removeAction(
-						existingAction.actionType,
-						event.guildId,
-						event.user.id
-					)
-				} else {
-					getLoggingChannelWithPerms(ConfigOptions.ACTION_LOG, event.getGuild())?.createEmbed {
-						title = "Banned a User"
-						description = "${event.user.mention} has been banned!"
-						baseModerationEmbed(event.getBan().reason, event.user, null)
-						timestamp = Clock.System.now()
-					}
-				}
-			}
-		}
-		event<BanRemoveEvent> {
-			check { anyGuild() }
-			action {
-				val ignore = ModerationActionCollection().shouldIgnoreAction(
-					ModerationAction.UNBAN,
-					event.guildId,
-					event.user.id
-				)
-				if (ignore != null && ignore) {
-					return@action
-				}
-
-				val existingAction =
-					ModerationActionCollection().getAction(ModerationAction.UNBAN, event.guildId, event.user.id)
-				if (existingAction != null && existingAction.targetUserId != event.user.id) {
-					return@action
-				}
-
-				if (existingAction != null) {
-					val isTempUnban = existingAction.data.reason?.contains("**temporary-ban**", true) == true
-
-					getLoggingChannelWithPerms(ConfigOptions.ACTION_LOG, event.getGuild())?.createEmbed {
-						title = if (isTempUnban) "Temporary ban removed" else "Unbanned a user"
-						description =
-							"${event.user.mention} has ${if (isTempUnban) "had their temporary ban removed!" else "been unbanned!"}\n${
-								event.user.id
-							}  (${event.user.username})"
-						if (existingAction.data.reason?.contains("**temporary-ban-expire**") == false) {
-							field {
-								name = "Reason:"
-								value = existingAction.data.reason.toString()
-							}
-						} else {
-							field {
-								name = "Initial Ban date:"
-								value = existingAction.data.timeData?.start?.toDiscord(TimestampType.ShortDateTime)
-									.toString()
-							}
-						}
-
-						if (existingAction.data.actioner != null) {
-							val user = UserBehavior(existingAction.data.actioner, kord).asUserOrNull()
-							footer {
-								text = user?.username ?: "Unable to get username"
-								icon = user?.avatar?.cdnUrl?.toUrl()
-							}
-						}
-						timestamp = Clock.System.now()
-						color = DISCORD_GREEN
-					}
-					ModerationActionCollection().removeAction(
-						existingAction.actionType,
-						event.guildId,
-						event.user.id
-					)
-				}
-			}
-		}
 		event<ChannelCreateEvent> {
 			check {
 				failIf {
@@ -224,7 +62,10 @@ class ModerationEvents : Extension() {
 				}
 			}
 			action {
-				val guild = event.channel.data.guildId.value?.let { GuildBehavior(it, event.kord) }
+				val guildId = event.channel.data.guildId.value
+				// Do not log if the channel logging option is false
+				if (guildId?.let { UtilityConfigCollection().getConfig(it)?.logChannelUpdates } == false) return@action
+				val guild = guildId?.let { GuildBehavior(it, event.kord) }
 				val perms = formatPermissionsForDisplay(guild, event.channel)
 				var allowed = perms.getValue(ALLOWED)
 				var denied = perms.getValue(DENIED)
@@ -252,7 +93,10 @@ class ModerationEvents : Extension() {
 		}
 		event<ChannelDeleteEvent> {
 			action {
-				val guild = event.channel.data.guildId.value?.let { GuildBehavior(it, event.kord) }
+				val guildId = event.channel.data.guildId.value
+				// Do not log if the channel logging option is false
+				if (guildId?.let { UtilityConfigCollection().getConfig(it)?.logChannelUpdates } == false) return@action
+				val guild = guildId?.let { GuildBehavior(it, event.kord) }
 				getLoggingChannelWithPerms(ConfigOptions.UTILITY_LOG, guild!!)?.createEmbed {
 					title = "${writeChannelType(event.channel.type)} Deleted"
 					description = "`${event.channel.data.name.value}` was deleted."
@@ -269,7 +113,10 @@ class ModerationEvents : Extension() {
 				}
 			}
 			action {
-				val guild = event.channel.data.guildId.value?.let { GuildBehavior(it, event.kord) }
+				val guildId = event.channel.data.guildId.value
+				// Do not log if the channel logging option is false
+				if (guildId?.let { UtilityConfigCollection().getConfig(it)?.logChannelUpdates } == false) return@action
+				val guild = guildId?.let { GuildBehavior(it, event.kord) }
 				val oldPerms = formatPermissionsForDisplay(guild, event.old)
 				val newPerms = formatPermissionsForDisplay(guild, event.channel)
 				val oldAllowed = oldPerms.getValue(ALLOWED)
@@ -433,6 +280,8 @@ class ModerationEvents : Extension() {
 		event<GuildScheduledEventCreateEvent> {
 			check { anyGuild() }
 			action {
+				// Do not log if event updates are disabled
+				if (UtilityConfigCollection().getConfig(event.guildId)?.logEventUpdates == false) return@action
 				val guild = GuildBehavior(event.guildId, kord)
 				getLoggingChannelWithPerms(ConfigOptions.UTILITY_LOG, guild)?.createEmbed {
 					title = "Scheduled Event Created!"
@@ -465,6 +314,8 @@ class ModerationEvents : Extension() {
 		event<GuildScheduledEventDeleteEvent> {
 			check { anyGuild() }
 			action {
+				// Do not log if event updates are disabled
+				if (UtilityConfigCollection().getConfig(event.guildId)?.logEventUpdates == false) return@action
 				val guild = GuildBehavior(event.guildId, kord)
 				getLoggingChannelWithPerms(ConfigOptions.UTILITY_LOG, guild)?.createEmbed {
 					title = "Scheduled Event Deleted!"
@@ -483,6 +334,8 @@ class ModerationEvents : Extension() {
 		event<GuildScheduledEventUpdateEvent> {
 			check { anyGuild() }
 			action {
+				// Do not log if event updates are disabled
+				if (UtilityConfigCollection().getConfig(event.guildId)?.logEventUpdates == false) return@action
 				val guild = GuildBehavior(event.guildId, kord)
 				getLoggingChannelWithPerms(ConfigOptions.UTILITY_LOG, guild)?.createEmbed {
 					title = "Scheduled Event Updated!"
@@ -527,6 +380,8 @@ class ModerationEvents : Extension() {
 		event<InviteCreateEvent> {
 			check { anyGuild() }
 			action {
+				// Do not log if invite updates are disabled
+				if (event.guildId?.let { UtilityConfigCollection().getConfig(it) }?.logInviteUpdates == false) return@action
 				val guild = event.guildId?.let { GuildBehavior(it, kord) } ?: return@action
 				getLoggingChannelWithPerms(ConfigOptions.UTILITY_LOG, guild)?.createEmbed {
 					title = "Invite link created"
@@ -563,6 +418,8 @@ class ModerationEvents : Extension() {
 		event<InviteDeleteEvent> {
 			check { anyGuild() }
 			action {
+				// Do not log if invite updates are disabled
+				if (event.guildId?.let { UtilityConfigCollection().getConfig(it) }?.logInviteUpdates == false) return@action
 				val guild = event.guildId?.let { GuildBehavior(it, kord) } ?: return@action
 				getLoggingChannelWithPerms(ConfigOptions.UTILITY_LOG, guild)?.createEmbed {
 					title = "Invite link deleted"
@@ -580,117 +437,10 @@ class ModerationEvents : Extension() {
 				}
 			}
 		}
-		event<MemberUpdateEvent> {
-			check { anyGuild() }
-			action {
-				val channel = getLoggingChannelWithPerms(ConfigOptions.ACTION_LOG, event.guild)
-				if (event.old?.timeoutUntil != event.member.timeoutUntil) {
-					var existingAction =
-						ModerationActionCollection().getAction(ModerationAction.TIMEOUT, event.guildId, event.member.id)
-
-					if (existingAction == null) {
-						existingAction =
-							ModerationActionCollection().getAction(
-								ModerationAction.REMOVE_TIMEOUT,
-								event.guildId,
-								event.member.id
-							)
-					}
-
-					if (existingAction != null) {
-						val data = existingAction.data
-						val isRemove = existingAction.actionType == ModerationAction.REMOVE_TIMEOUT
-						val user = UserBehavior(existingAction.targetUserId, kord).asUserOrNull()
-						val actioner = data.actioner?.let { UserBehavior(it, kord) }?.asUserOrNull()
-						channel?.createEmbed {
-							if (isRemove) {
-								title = "Timeout removed"
-								dmNotificationStatusEmbedField(data.dmOutcome, data.dmOverride)
-								field {
-									name = "User:"
-									value = "${user?.username} (${user?.id})"
-								}
-								footer {
-									text = "Requested by ${actioner?.username}"
-									icon = user?.avatar?.cdnUrl?.toUrl()
-								}
-								timestamp = Clock.System.now()
-								color = DISCORD_GREEN
-							} else {
-								title = "Timeout"
-								image = data.imageUrl
-								baseModerationEmbed(data.reason, event.member, user)
-								dmNotificationStatusEmbedField(data.dmOutcome, data.dmOverride)
-								timestamp = data.timeData?.start
-								color = DISCORD_RED
-								field {
-									name = "Duration"
-									value = data.timeData?.durationInst?.toDiscord(TimestampType.Default) +
-										" (${data.timeData?.durationDtp.interval()})"
-								}
-							}
-						}
-
-						ModerationActionCollection().removeAction(existingAction.actionType, event.guild.id, event.member.id)
-					} else {
-						channel?.createEmbed {
-							if (event.member.timeoutUntil != null) {
-								title = "Timeout"
-								color = DISCORD_RED
-								field {
-									name = "Duration"
-									value = event.member.timeoutUntil?.toDiscord(TimestampType.Default) ?: "0"
-								}
-							} else {
-								title = "Removed timeout"
-								color = DISCORD_GREEN
-							}
-							field {
-								name = "User"
-								value = event.member.mention + "(${event.member.id})"
-							}
-							description = "via Discord menus"
-							timestamp = Clock.System.now()
-						}
-					}
-				} else {
-						val newRoles = mutableListOf<String>()
-						val oldRoles = mutableListOf<String>()
-						event.member.roleBehaviors.forEach { newRoles.add(it.mention) }
-						event.old?.roleBehaviors?.forEach { oldRoles.add(it.mention) }
-						channel?.createEmbed {
-							title = "Member updated"
-							description = "${event.member.username} has been updated"
-							if (event.member.nickname != event.old?.nickname) {
-								field {
-									name = "Nickname change"
-									value = "Old: ${event.old?.nickname}\nNew: ${event.member.nickname}"
-								}
-							}
-							if (event.member.isOwner() != event.old?.isOwner()) {
-								field {
-									name = "Became server owner"
-								}
-							}
-							if (event.member.roles != event.old?.roles) {
-								field {
-									name = "New Roles"
-									value = newRoles.joinToString(", ")
-									inline = true
-								}
-								field {
-									name = "Old roles"
-									value = oldRoles.joinToString(", ")
-									inline = true
-								}
-							}
-							color = DISCORD_YELLOW
-						}
-					}
-				}
-		}
 		event<RoleCreateEvent> {
 			action {
+				// Do not log if role updates are disabled
+				if (UtilityConfigCollection().getConfig(event.guildId)?.logRoleUpdates == false) return@action
 				val guild = GuildBehavior(event.guildId, kord)
 				getLoggingChannelWithPerms(ConfigOptions.UTILITY_LOG, guild)?.createEmbed {
 					title = "Created a Role"
@@ -746,6 +496,8 @@ class ModerationEvents : Extension() {
 		}
 		event<RoleDeleteEvent> {
 			action {
+				// Do not log if role updates are disabled
+				if (UtilityConfigCollection().getConfig(event.guildId)?.logRoleUpdates == false) return@action
 				val guild = GuildBehavior(event.guildId, kord)
 				getLoggingChannelWithPerms(ConfigOptions.UTILITY_LOG, guild)?.createEmbed {
 					title = "Role deleted"
@@ -761,6 +513,8 @@ class ModerationEvents : Extension() {
 		}
 		event<RoleUpdateEvent> {
 			action {
+				// Do not log if role updates are disabled
+				if (UtilityConfigCollection().getConfig(event.guildId)?.logRoleUpdates == false) return@action
 				val guild = GuildBehavior(event.guildId, kord)
 				val channel = getLoggingChannelWithPerms(ConfigOptions.UTILITY_LOG, guild)
 				channel?.createMessage {
@@ -832,6 +586,8 @@ class ModerationEvents : Extension() {
 		event<ThreadChannelCreateEvent> {
 			check { anyGuild() }
 			action {
+				// Do not log if the channel logging option is false
+				if (UtilityConfigCollection().getConfig(event.channel.guild.id)?.logChannelUpdates == false) return@action
 				val appliedTags = mutableListOf<String>()
 				event.channel.appliedTags.forEach { tag ->
 					event.channel.parent.asChannelOfOrNull<ForumChannel>()
@@ -863,6 +619,8 @@ class ModerationEvents : Extension() {
 		event<ThreadChannelDeleteEvent> {
 			check { anyGuild() }
 			action {
+				// Do not log if the channel logging option is false
+				if (UtilityConfigCollection().getConfig(event.channel.guild.id)?.logChannelUpdates == false) return@action
 				getLoggingChannelWithPerms(ConfigOptions.UTILITY_LOG, event.channel.guild)?.createEmbed {
 					title = "${writeChannelType(event.channel.type)} Deleted"
 					description = "`${event.channel.data.name.value}` was deleted."
